requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.5
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.4
django-allauth==0.50.0
django-cleanup==6.0.0
django-cors-headers==3.12.0
django-extensions==3.1.5
django-uuidfield==0.5.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.15.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.4
oauthlib==3.2.0
Pillow==9.2.0
prompt-toolkit==3.0.30
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
